requirements.txt
openai
requests
python-dotenv
twilio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib